@@ -2,7 +2,7 @@
   "name": "netbrothers-gmbh/nbcsb-bundle",
   "description": "Creating skeleton structure for symfony bundle with one single command. Creates directory structure and necessary files, activates your bundle. See: https://netbrothers.de/easy-create-reusable-symfony-bundle",
   "type": "symfony-bundle",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "keywords": ["symfony","bundle","helper","symfony command line"],
   "homepage": "https://github.com/netbrothers-gmbh/nbcsb-bundle",
   "license": "MIT",
@@ -15,7 +15,7 @@
     }
   ],
   "require": {
-    "php": ">=7.2.5",
+    "php": ">=8.1.0",
     "symfony/console": "5.*",
     "symfony/serializer": "5.*",
     "symfony/string": "5.*",
@@ -24,11 +24,6 @@
   "autoload": {
     "psr-4": {
       "NetBrothers\\NbCsbBundle\\": "src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "NetBrothers\\NbCsbBundle\\Tests\\": "tests/"
     }
   }
 }
