@@ -40,7 +40,7 @@ class NetBrothersNbCsbExtension extends Extension
     /**
      * @return string
      */
-    public function getAlias()
+    public function getAlias(): string
     {
         return 'netbrothers_nbcsb';
     }
