@@ -9,6 +9,7 @@
 
 namespace NetBrothers\NbCsbBundle;
 use NetBrothers\NbCsbBundle\DependencyInjection\NetBrothersNbCsbExtension;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class NetBrothersNbCsbBundle extends Bundle
@@ -16,7 +17,7 @@ class NetBrothersNbCsbBundle extends Bundle
     /**
      * Overridden to allow for the custom extension alias.
      */
-    public function getContainerExtension()
+    public function getContainerExtension(): ?ExtensionInterface
     {
         if (null === $this->extension) {
             $this->extension = new NetBrothersNbCsbExtension();
